+++ src/scan1blup.h
@@ -13,7 +13,6 @@
 // se        = If TRUE, calculate SEs
 // reml      = If TRUE, use REML to estimate variance components; otherwise use maximum
 //             likelihood
-// preserve_intercept = If FALSE, add the intercept to the BLUPs and remove that column
 // tol       = Numeric tolerance
 //
 // output    = List with two matrices, of coefficients and SEs (each coefficients x positions)
@@ -22,7 +21,6 @@ Rcpp::List scanblup(const Rcpp::NumericVector& genoprobs,
                     const Rcpp::NumericMatrix& addcovar,
                     const bool se,
                     const bool reml,
-                    const bool preserve_intercept,
                     const double tol);
 
 #endif // SCAN1BLUP_H